0.0 to 0

+++ datasets/Moran_2021_corrected/compose.py
@@ -13,8 +13,8 @@ df_labels = df_labels.reindex(columns=["title", "label_included", "label_abstrac
 df = pd.merge(df_abstracts, df_labels, how="outer", on="title")
 
 df["doi"] = "https://doi.org/" + df["url"].str.extract(r"(10\.[^&]*)")[0]
-df['label_included'] = df['label_included'].fillna(0)
-df['label_abstract_included'] = df['label_abstract_included'].fillna(0)
+df['label_included'] = df['label_included'].fillna(0).astype(int)
+df['label_abstract_included'] = df['label_abstract_included'].fillna(0).astype(int)
 
 # save results to file
 df.to_csv(f"{key}_raw.csv", index=False)
